Add : mainType 파일 생성 후 각각 페이지 타입 정리
