+++ src/api/firebase.tsx
@@ -1,5 +1,7 @@
 import { initializeApp } from "firebase/app";
 import { v4 as uuidv4 } from "uuid";
+import { SetterOrUpdater } from "recoil";
+
 import {
   getAuth,
   signInWithPopup,
@@ -8,6 +10,7 @@ import {
   signOut,
   onAuthStateChanged,
 } from "firebase/auth";
+
 import {
   getDatabase,
   ref,
@@ -21,6 +24,7 @@ import {
   remove,
   update,
 } from "firebase/database";
+
 import {
   UsedItemType,
   UserDataType,
@@ -29,91 +33,19 @@ import {
   MessagesType,
   MessageListType,
 } from "../types/usedType";
-import { SetterOrUpdater } from "recoil";
+
+import {
+  Product,
+  AddProduct,
+  PayProduct,
+  GetOrderDetails,
+  OrderDetails,
+  Comment,
+  CartProducts,
+} from "../types/mainType";
+
 interface AdminUser extends User {
   isAdmin: boolean;
-}
-
-export interface Product {
-  id: string;
-  title: string;
-  category: string;
-  description: string;
-  price: string;
-  image: string;
-  options: string[];
-}
-
-export interface addProduct {
-  id: string;
-  title: string;
-  category: string;
-  description: string;
-  price: string;
-  image: string;
-  options: string;
-}
-
-export interface PayProduct extends Product {
-  quantity: number;
-}
-
-export interface testOrderProduct {
-  title: string;
-  description: string;
-  price: string;
-  image: string;
-  options: string;
-  quantity: number;
-}
-
-export interface testProduct {
-  title: string;
-  description: string;
-  price: string;
-  image: string;
-  options: string[];
-  quantity: number;
-}
-
-export interface getOrderDetails {
-  ordersId?: string;
-  name: string;
-  phone: string;
-  address: string;
-  paymentMethod: string;
-  createdAt?: string;
-  items: testProduct[];
-}
-
-export interface OrderDetails {
-  ordersId?: string;
-  name: string;
-  phone: string;
-  address: string;
-  paymentMethod: string;
-  createdAt?: string;
-}
-
-interface Comment {
-  id: string;
-  text: string;
-  rank: number;
-  createdAt: string;
-  uid: string;
-  userPhoto: string;
-  displayName: string;
-}
-
-export interface CartProducts {
-  id: string;
-  title: string;
-  category: string;
-  description: string;
-  price: string;
-  image: string;
-  options: string;
-  quantity: number;
 }
 
 const firebaseConfig = {
@@ -210,7 +142,7 @@ export async function getProducts(): Promise<Product[]> {
 }
 
 export async function addNewProduct(
-  product: addProduct,
+  product: AddProduct,
   image: string
 ): Promise<void> {
   const id = uuidv4();
@@ -221,10 +153,8 @@ export async function addNewProduct(
     id,
     price: product.price,
     image,
-    // options: product.options,
     options: product.options.split(","),
   });
-  // return;
 }
 
 export async function addWishlistItem(
@@ -241,13 +171,6 @@ export async function setWishlistItems(
 ): Promise<void> {
   await set(ref(database, `wishlist/${userId}`), wishlist);
 }
-
-// export async function removeWishlistItem(
-//   userId: string,
-//   product: Product
-// ): Promise<void> {
-//   return remove(ref(database, `wishlist/${userId}/${product.id}}`));
-// }
 
 export async function getWishlistItems(userId: string): Promise<Product[]> {
   const wishlistRef = ref(database, `wishlist/${userId}`);
@@ -324,7 +247,7 @@ export async function addOrderList(
 
 export async function getOrderItems(
   userId: string
-): Promise<getOrderDetails[]> {
+): Promise<GetOrderDetails[]> {
   const orderItemsRef = ref(getDatabase(), `userData/${userId}/orders`);
   return get(orderItemsRef).then((snapshot) => {
     if (snapshot.exists()) {
